MySql file is added only with table 'user', it will keep on changing
This file is just for import purpose to see how the table for users
would look like. Collaborators should look at this file to generate
their tables. You have to do all of the database work for your task
for yourself and at the end we would make it into and put on the
server.